+++ components/flight-pagination.tsx
@@ -1,0 +1,55 @@
+"use client"
+
+import type { Dispatch, SetStateAction } from "react"
+
+interface FlightPaginationProps {
+  currentPage: number
+  totalPages: number
+  setCurrentPage: Dispatch<SetStateAction<number>>
+}
+
+export default function FlightPagination({ currentPage, totalPages, setCurrentPage }: FlightPaginationProps) {
+  if (totalPages <= 1) return null
+
+  return (
+    <div className="flex justify-center mt-6">
+      <div className="flex items-center space-x-2">
+        <button
+          onClick={() => setCurrentPage((prev) => Math.max(prev - 1, 1))}
+          disabled={currentPage === 1}
+          className={`px-3 py-1 rounded-md ${
+            currentPage === 1
+              ? "bg-gray-100 text-gray-400 cursor-not-allowed"
+              : "bg-white border border-gray-200 text-gray-700 hover:bg-gray-50"
+          }`}
+        >
+          Previous
+        </button>
+        {Array.from({ length: totalPages }, (_, i) => i + 1).map((page) => (
+          <button
+            key={page}
+            onClick={() => setCurrentPage(page)}
+            className={`px-3 py-1 rounded-md ${
+              currentPage === page
+                ? "bg-black text-white"
+                : "bg-white border border-gray-200 text-gray-700 hover:bg-gray-50"
+            }`}
+          >
+            {page}
+          </button>
+        ))}
+        <button
+          onClick={() => setCurrentPage((prev) => Math.min(prev + 1, totalPages))}
+          disabled={currentPage === totalPages}
+          className={`px-3 py-1 rounded-md ${
+            currentPage === totalPages
+              ? "bg-gray-100 text-gray-400 cursor-not-allowed"
+              : "bg-white border border-gray-200 text-gray-700 hover:bg-gray-50"
+          }`}
+        >
+          Next
+        </button>
+      </div>
+    </div>
+  )
+}
